requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-annotate requirements.in
#
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.5
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
slackclient==2.5.0
urllib3==1.25.9
yarl==1.4.2
